Struct string::String

pub struct String<T = Vec<u8>> { /* fields omitted */ }

A UTF-8 encoded string with configurable byte storage.

This type differs from std::String in that it is generic over the underlying byte storage, enabling it to use Vec<[u8]>, &[u8], or third party types, such as Bytes.

In order to construct String via any of the non-unsafe constructors, the backing storage needs to implement the StableAsRef marker trait. If you wish to construct String with a type that does not implement StableAsRef, you can use the from_utf8_unchecked constructor.

Implementations

impl<T> String<T>

pub fn get_ref(&self) -> &T

Get a reference to the underlying byte storage.

Examples

let s = String::new();
let vec = s.get_ref();

pub unsafe fn get_mut(&mut self) -> &mut T

Get a mutable reference to the underlying byte storage.

It is inadvisable to directly manipulate the byte storage. This function is unsafe as the bytes could no longer be valid UTF-8 after mutation.

Examples

let mut s = String::new();

unsafe {
    let vec = s.get_mut();
}

pub fn into_inner(self) -> T

Unwraps this String, returning the underlying byte storage.

Examples

let s = String::new();
let vec = s.into_inner();

pub fn from_str<'a>(src: &'a str) -> String<T> where
T: From<&'a [u8]> + StableAsRef,

Creates a new String from a &str.

Use TryFrom for conversion from &u8.

let _: String<Vec<u8>> = String::from_str("nice str");

impl String

pub fn new() -> String

Creates a new empty String.

Given that the String is empty, this will not allocate.

Examples

Basic usage

let s = String::new();
assert_eq!(s, "");

impl<T> String<T> where
T: AsRef<[u8]>,

pub unsafe fn from_utf8_unchecked(value: T) -> String<T>

Converts the provided value to a String without checking that the given value is valid UTF-8.

Use TryFrom for a safe conversion.

Safety

You must ensure that:

The backing storage type T adheres to the contract as documented on the StableAsRef marker trait.

If T implements AsRef<[u8]> and/or AsMut<[u8]>, the byte slice returned by calling as_ref and/or as_mut on the provided value represents valid utf-8.